Capture: event-set sub-component Fixes #44
