--- src/app/features/components/capture/current-event/begining-type-setting/begining-type-setting.component.html
@@ -1,8 +1,0 @@
-<mat-radio-group
-  aria-labelledby="radio-buttons-group-label"
-  class="radio-buttons-group"
-  [(ngModel)]="selectedEventType"  (change)="onSelectionChange($event)">
-  @for (eventType of eventTypes; track eventType) {
-    <mat-radio-button class="radio-button" [value]="eventType">{{eventType}}...</mat-radio-button>
-  }
-</mat-radio-group>
