--- src/app/features/components/capture/current-event/begining-type-setting/begining-type-setting.component.scss
@@ -1,1 +1,0 @@
-@import '../../../../../shared/styles/radio-buttons.scss';
