+++ src/app/features/components/capture/current-event/beginning-type-setting/beginning-type-setting.component.scss
@@ -1,0 +1,1 @@
+@import '../../../../../shared/styles/radio-buttons.scss';
