+++ src/app/features/components/capture/current-event/current-event.component.html
@@ -11,8 +11,8 @@
             <app-event-type-setting></app-event-type-setting>
         } 
 
-        @case ("begining-type-setting") {
-            <app-begining-type-setting></app-begining-type-setting>
+        @case ("beginning-type-setting") {
+            <app-beginning-type-setting></app-beginning-type-setting>
         }
 
         @case ("ressource-type-setting") {
@@ -47,9 +47,6 @@
             <app-interval-type-setting></app-interval-type-setting>
         }
 
-        @case ("beginning-type-setting") {
-            <app-beginning-type-setting></app-beginning-type-setting>
-        }
 
        @case ("period-type-setting") {
            <app-period-type-setting></app-period-type-setting>
