+++ src/app/features/components/capture/current-event/period-type-setting/period-type-setting.component.scss
@@ -1,0 +1,1 @@
+@import '../../../../../shared/styles/radio-buttons.scss';
